i/prompting/patterns: consolidate render node types into render.go
